+++ src/globals.cpp
@@ -2,10 +2,10 @@
 
 // ============ 全局变量定义 ============
 int random_val;
-std::map<float, std::string> rooms;
+std::map<int, std::string> rooms;
 std::map<int, char> alph;
 std::map<int, std::string> dire;
-float Prooms;
+int Prooms;
 
 bool chap1 = true, chap2 = true, chap3 = true;
 std::string Chap1 = "在起初，地狱乏火与死之人间";
